requirements-dev.txt
aiohttp>=2.3.0,<=3.2.1
sanic-testing>=0.8.2
nose
